requirements.txt
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
git+https://github.com/Rapptz/discord.py@rewrite
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
=======
git+https://github.com/Rapptz/discord.py.git
dnspython
PyNaCl
praw
async-timeout
>>>>>>> 7ea5e90 (Update Readme)
=======
discord.py[voice]
<<<<<<< HEAD
<<<<<<< HEAD
>>>>>>> 0c218fc (update requirements to use discord.py with voice support)
=======
discord.py[voice]
python-dotenv
>>>>>>> f90b46c (Add python-dotenv to requirements for environment variable management)
=======
python-dotenv
mkdocs-material
mkdocs-awesome-pages-plugin
<<<<<<< HEAD
<<<<<<< HEAD
>>>>>>> b68b48d (delete procfile, add feature to mkdocs.yml, added ci.yml to deploy to github pages)
=======
pymdown-extensions
>>>>>>> 7f228e7 (Add pymdown-extensions to requirements for enhanced Markdown support)
=======
discord.py[voice]
>>>>>>> 0c218fc (update requirements to use discord.py with voice support)
=======
discord.py[voice]
python-dotenv
>>>>>>> f90b46c (Add python-dotenv to requirements for environment variable management)
=======
python-dotenv
mkdocs-material
mkdocs-awesome-pages-plugin
>>>>>>> b68b48d (delete procfile, add feature to mkdocs.yml, added ci.yml to deploy to github pages)
=======
pymdown-extensions
>>>>>>> 7f228e7 (Add pymdown-extensions to requirements for enhanced Markdown support)
